day2prgm1.c

+++ day2prgm1.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-long long factorial(int n) {
+long factorial(int n) {
     if (n < 0) {
         printf("Factorial is not defined for negative numbers.\n");
         return -1;
